refactor data processing
